move implementation base class to separate module
